update file upload.

+++ plugins/file_uploader/lib/file_uploader.dart
@@ -8,6 +8,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart' show WidgetsBindingObserver, AppLifecycleState, WidgetsBinding;
 import 'package:uuid/uuid.dart';
 
+import 'exception.dart';
 import 'http_upload.dart';
 import 'store.dart';
 import 'task.dart';
@@ -38,12 +39,39 @@ void _innerPrint(str) {
   }
 }
 
+class FileUploaderFinishResponse {
+  final String taskId;
+  final bool success;
+  FileUploaderFinishResponse(this.taskId, this.success);
+
+  @override
+  String toString() {
+    return '{taskId: $taskId, success: $success}';
+  }
+}
+
+class FileUploaderProgressResponse {
+  final String taskId;
+  final num progress;
+  FileUploaderProgressResponse({this.taskId, this.progress});
+
+  @override
+  String toString() {
+    return '{taskId: $taskId, progress: $progress}';
+  }
+}
+
 class FileUploader {
   static const MethodChannel _channel = const MethodChannel('file_uploader');
 
   /// 全局uploader对象
 //  static var _uploader = FlutterUploader();
   static var _taskSession = UploadTaskSession()..init();
+  static var _finishController = StreamController();
+  static var _progressController = StreamController();
+
+  static Stream get finishStream => _finishController.stream.asBroadcastStream();
+  static Stream get progressStream => _progressController.stream.asBroadcastStream();
 
   static Future<String> get platformVersion async {
     final String version = await _channel.invokeMethod('getPlatformVersion');
@@ -66,7 +94,29 @@ class FileUploader {
     _taskSession.addTask(null);
   }
 
-  static Future<String> enqueue(List<String> filePaths, {Map<String, dynamic> headers = const {}, Map<String, String> fields = const {}}) async {
+  static String _copyTmpFile(List<String> filePaths, String taskId) {
+    String newDir = '${getCacheDir()}/${taskId}';
+    if (!Directory(newDir).existsSync()) {
+      Directory(newDir).createSync(recursive: true);
+    }
+    int copyCount = 0;
+    for (var curPath in filePaths) {
+      if (File(curPath).existsSync()) {
+        var fileName = curPath.split('/').last;
+        File(curPath).copySync('${newDir}/${fileName}');
+        copyCount++;
+      }
+    }
+    _innerPrint('copyCount = ${copyCount}');
+    return newDir;
+  }
+
+  static Future<String> enqueue(
+    List<String> filePaths, {
+    Map<String, dynamic> headers = const {},
+    Map<String, String> fields = const {},
+    bool immediately = false,
+  }) async {
     Map<String, String> temps = {};
     headers.forEach((str, value) {
       temps.putIfAbsent(str, () => value.toString());
@@ -75,24 +125,65 @@ class FileUploader {
     commonFields = fields;
     var taskId = uuid.v4();
     var dataStr = '{}';
-    if (_config.archived) {
-      var newDir = '${getCacheDir()}/${taskId}';
-      _innerPrint('newDir = ${newDir}');
-      if (!Directory(newDir).existsSync()) {
-        Directory(newDir).createSync(recursive: true);
-      }
-      for (var curPath in filePaths) {
-        if (File(curPath).existsSync()) {
-          var fileName = curPath.split('/').last;
-          File(curPath).copySync('${newDir}/${fileName}');
+    if (immediately) {
+      Function asyncFun = () async {
+        var paramsStruct = TaskParamsStruct(
+          fields: fields,
+          filePaths: filePaths,
+          archived: _config.archived ?? true,
+          immediately: true,
+          taskId: taskId,
+        );
+
+        if (_config.archived) {
+          String newDir = _copyTmpFile(filePaths, taskId);
+          paramsStruct = TaskParamsStruct(
+            fields: fields,
+            dirPath: newDir,
+            archived: _config.archived ?? true,
+            immediately: true,
+            taskId: taskId,
+          );
         }
-      }
-      dataStr = TaskParamsStruct(fields: fields, dirPath: newDir, filePaths: [], archived: _config.archived ?? true).toStoreString();
+        bool isWifi = await connect.checkConnectivity() == ConnectivityResult.wifi;
+        paramsStruct.config = _config;
+        paramsStruct.isWifi = isWifi;
+        paramsStruct.headers = commonHeaders;
+        var response = UploadResponse(errCode: -1);
+        try {
+          response = await executeTask(paramsStruct);
+        } catch (e) {
+          if (checkExceptionValid(e)) {
+            response = UploadResponse(errCode: 0);
+          }
+          print('执行任务异常:$e');
+        }
+
+        _finishController.sink.add(FileUploaderFinishResponse(taskId, response.errCode == 0));
+      };
+      asyncFun();
+    } else if (_config.archived) {
+      String newDir = _copyTmpFile(filePaths, taskId);
+      dataStr = TaskParamsStruct(
+        fields: fields,
+        dirPath: newDir,
+        filePaths: [],
+        archived: _config.archived ?? true,
+        taskId: taskId,
+      ).toStoreString();
+      _taskSession.addTask(UploadTaskItem(taskId: taskId, data: dataStr));
     } else {
-      dataStr = TaskParamsStruct(fields: fields, dirPath: '', filePaths: filePaths, archived: _config.archived ?? true).toStoreString();
+      dataStr = TaskParamsStruct(
+        fields: fields,
+        dirPath: '',
+        filePaths: filePaths,
+        archived: _config.archived ?? true,
+        taskId: taskId,
+      ).toStoreString();
+      _taskSession.addTask(UploadTaskItem(taskId: taskId, data: dataStr));
     }
-    _taskSession.addTask(UploadTaskItem(taskId: taskId, data: dataStr));
-    return '';
+
+    return taskId;
   }
 }
 
@@ -145,7 +236,7 @@ class UploadTaskSession extends TaskSession<UploadTaskItem> with WidgetsBindingO
       List items = await UploadTaskProvider.getItems(limit: 1);
       if (items.length > 0) {
         var item0 = items[0];
-        print('item0 = ${item0}');
+        _innerPrint('item0 = ${item0}');
         try {
           bool isWifi = await connect.checkConnectivity() == ConnectivityResult.wifi;
           TaskParamsStruct paramsStruct = TaskParamsStruct.fromStoreString(item0.data);
@@ -153,7 +244,7 @@ class UploadTaskSession extends TaskSession<UploadTaskItem> with WidgetsBindingO
           paramsStruct.isWifi = isWifi;
           paramsStruct.headers = commonHeaders;
 
-          var uploadRsp = await compute(executeTask, paramsStruct);
+          var uploadRsp = await executeTask(paramsStruct);
           if (uploadRsp.errCode == 0) {
             await removeTask(item0.taskId);
           } else {
@@ -161,16 +252,9 @@ class UploadTaskSession extends TaskSession<UploadTaskItem> with WidgetsBindingO
             await Future.delayed(Duration(seconds: 20));
           }
         } catch (e) {
-          if (e is Exception) {
-            /// 处理 Exception: Invalid argument(s): Illegal argument in isolate message : (object is a closure - Function '<anonymous closure>':.)
-            /// 其实属于上传成功
-            /// var invalidMessagePre = 'Exception: Invalid argument(s): Illegal argument in isolate message';
-            /// if (e.toString().startsWith(invalidMessagePre)) {
-            ///  print('处理 Exception: Invalid argument(s): Illega');
-            /// }
-          }
+          if (checkExceptionValid(e)) {}
           await removeTask(item0.taskId);
-          print('getNextTask $e');
+          _innerPrint('getNextTask $e');
         }
       }
     };
@@ -190,7 +274,8 @@ Future<UploadResponse> executeTask(TaskParamsStruct params) async {
 
   final config = params.config;
   var filePaths = params.filePaths;
-  _innerPrint('executeTask: ${params.filePaths}');
+  _innerPrint('executeTask: filePaths = ${params.filePaths}');
+  _innerPrint('executeTask: dirPath = ${params.dirPath}');
   var userId = 'unknown';
   if (params.fields != null && params.fields['userId'] is String) {
     userId = params.fields['userId'];
@@ -201,25 +286,50 @@ Future<UploadResponse> executeTask(TaskParamsStruct params) async {
   var platform = Platform.isIOS ? 'ios' : 'android';
   var newPath = '${Directory.systemTemp.path}/${platform}_${userId}_${timeStr}.zip';
   if (params.archived) {
-    var filePath = await zipEncoderDir(params.dirPath);
+    var filePath = await compute(zipEncoderDir, params.dirPath);
     if (filePath != null) {
       // 这里重命名文件
       var file = File(filePath);
-      file.rename(newPath);
+      file.renameSync(newPath);
       filePaths = [newPath];
+      _innerPrint('重命名成功');
+    } else {
+      _innerPrint('失败成功');
     }
   }
 
-  var res = await uploadFile(params.config.url, filePaths, fields: params.fields, headers: params.headers);
+  StreamController controller = StreamController();
+  controller.stream.listen((progress) {
+    FileUploader._progressController.sink.add(FileUploaderProgressResponse(taskId: params.taskId, progress: progress));
+  });
+  var res = await uploadFile(
+    params.config.url,
+    filePaths,
+    fields: params.fields,
+    headers: params.headers,
+    progressController: controller,
+  );
+  _innerPrint('uploadFile = $res');
   try {
     if (config.archived && res.errCode == 0) {
       if (File(newPath).existsSync()) {
         File(newPath).deleteSync(recursive: true);
       }
 
-      if (Directory(params.dirPath).existsSync()) {
+      if (params.dirPath != null && Directory(params.dirPath).existsSync()) {
         Directory(params.dirPath).deleteSync(recursive: true);
       }
+
+      _innerPrint('删除缓存数据成功');
+    } else if (params?.immediately ?? false) {
+      if (File(newPath).existsSync()) {
+        File(newPath).deleteSync(recursive: true);
+      }
+
+      if (params.dirPath != null && Directory(params.dirPath).existsSync()) {
+        Directory(params.dirPath).deleteSync(recursive: true);
+      }
+      _innerPrint('删除缓存数据成功');
     }
   } catch (e) {
     print('delete upload file: $e');
@@ -230,6 +340,7 @@ Future<UploadResponse> executeTask(TaskParamsStruct params) async {
 
 /// 执行任务的参数结构体。因为不同的 isolate 不能共享参数。
 class TaskParamsStruct {
+  final String taskId;
   final List<dynamic> filePaths; // store
   final String dirPath; // store
   final Map<String, String> fields; // store
@@ -237,8 +348,19 @@ class TaskParamsStruct {
   Map<String, String> headers;
   FileUploaderConfig config;
   bool isWifi;
+  bool immediately;
 
-  TaskParamsStruct({this.filePaths, this.headers, this.fields, this.config, this.isWifi = false, this.dirPath, this.archived});
+  TaskParamsStruct({
+    this.taskId,
+    this.filePaths,
+    this.headers,
+    this.fields,
+    this.config,
+    this.isWifi = false,
+    this.dirPath,
+    this.archived,
+    this.immediately = false,
+  });
 
   String toStoreString() {
     var map = {
@@ -246,6 +368,7 @@ class TaskParamsStruct {
       'dirPath': dirPath,
       'fields': fields,
       'archived': archived,
+      'taskId': taskId.toString(),
     };
 
     return jsonEncode(map);
@@ -254,7 +377,13 @@ class TaskParamsStruct {
   static TaskParamsStruct fromStoreString(String string) {
     try {
       var map = jsonDecode(string);
-      return TaskParamsStruct(filePaths: map['filePaths'], dirPath: map['dirPath'], fields: Map<String, String>.from(map['fields']), archived: map['archived']);
+      return TaskParamsStruct(
+        filePaths: map['filePaths'],
+        dirPath: map['dirPath'],
+        fields: Map<String, String>.from(map['fields']),
+        archived: map['archived'],
+        taskId: map['taskId'],
+      );
     } catch (e) {
       print('fromStoreString fail:$e');
     }
